% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{bootstrap_df}
\alias{bootstrap_df}
\title{Bootstrap an AFQ dataframe}
\usage{
bootstrap_df(
  input_df,
  dwi_metric,
  group.by = "group",
  participant.id = "subjectID"
)
}
\arguments{
\item{input_df}{The input AFQ dataframe}

\item{dwi_metric}{The diffusion MRI metric (e.g. "FA", "MD")}

\item{group.by}{The grouping variable used to group nodeID smoothing terms}

\item{participant.id}{The name of the column that encodes participant ID}
}
\value{
A shuffled AFQ dataframe
}
\description{
This function bootstrap samples an AFQ dataframe by participant.
That is, it first pivots to wide format with one row per participant,
bootstrap samples, and finally pivots back to long format.
}
\examples{
\dontrun{
df_afq <- read.csv("/path/to/afq/output.csv")
df_boot <- bootstrap_df(df_afq, "dti_fa")
}
}
