% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{estimate_distribution}
\alias{estimate_distribution}
\title{Estimate distribution function.}
\usage{
estimate_distribution(
  x,
  distr_names = c("beta", "gamma", "norm"),
  eval_metric = "aic"
)
}
\arguments{
\item{x}{A numeric vector that will be evaluated.}

\item{eval_metric}{The distribution evaluation metric names. \cr
Possible options: {'aic', 'bic', 'loglik'} \cr
Default: "aic"}
}
\value{
The best fitting family function to the values.
}
\description{
Estimate distribution function.
}
\details{
See \code{\link[fitdistrplus:fitdist]{fitdistrplus::fitdist()}} for additional information.
}
\examples{
x <- rnorm(1000)
estimate_distribution(x)

x <- rgamma(1000, shape = 12)
estimate_distribution(x)
}
