% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{estimate_smooth_basis}
\alias{estimate_smooth_basis}
\title{Estimate smoothing basis dimensions for GAM smoothers.}
\usage{
estimate_smooth_basis(target, df, regressors = NULL, family = "auto", ...)
}
\arguments{
\item{target}{The column name that encodes the metric to model.}

\item{regressors}{Column name or list of column names to use as
regressors, not including nodes smoothing terms and
the participant random effect. This list can also
include smoothing terms. Default: NULL.}

\item{family}{The family}
}
\value{
description
}
\description{
Estimate smoothing basis dimensions for GAM smoothers.
}
