% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgam.R
\name{fit_gam}
\alias{fit_gam}
\title{Fit a GAM for tract node metrics (e.g. FA, MD)}
\usage{
fit_gam(
  df_tract,
  target,
  covariates,
  k = 40,
  family = "auto",
  tract_name = "",
  out_dir = ".",
  save_output = FALSE
)
}
\arguments{
\item{df_tract}{AFQ Dataframe of node metric values for single tract}

\item{target}{The diffusion metric to model (e.g. "FA", "MD")}

\item{covariates}{List of strings of GAM covariates, not including
the smoothing terms over nodes and the random effect due to subjectID.
This list can also include smoothing terms.}

\item{k}{Dimension of the basis used to represent the node smoothing term,
If k = 'auto', this function will attempt to find the best value}

\item{family}{Distribution to use for the gam. Must be either 'gamma', 'beta', or 'auto'.}

\item{tract_name}{Name of the tract, used only for output file names}

\item{out_dir}{Directory in which to save gam stats}

\item{save_output}{Boolean flag to save gam stat files}
}
\value{
Fitted GAM model
}
\description{
This function has a series of steps:
\itemize{
\item If family == "auto", choose the distribution (either beta or gamma)
that has the lowest AIC when fitting to the dMRI metric data
\item If k == "auto", build an initial GAM model with k = 16 and continue to
double the k value until gam.check shows that k is large enough
\item Fit a GAM model such that: \cr \cr
target ~ covariates + s(nodeID, by=group, k = k_value) + s(subjectID, bs = "re")
\cr
\item Optionally save the output of gam.check and summary to files.
}
}
\examples{
\dontrun{
df_afq <- read.csv("/path/to/afq/output.csv")
gam_fit <- fit_gam(df_afq,
                   target = "dti_fa",
                   covariates = list("group", "sex"),
                   family = "gamma",
                   k = "auto")
}
}
