% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load.afq.sarica}
\alias{load.afq.sarica}
\title{Load tract profiles from Sarica et al.}
\usage{
load.afq.sarica(...)
}
\arguments{
\item{...}{arguments to be passed to read.csv}
}
\value{
A merged dataframe with data from Sarica et al.
}
\description{
Load tract profiles from Sarica et al.
}
\examples{
df_sarica <- load.afq.sarica()
}
