% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{parse_s3_uri}
\alias{parse_s3_uri}
\title{Parse an S3 URI}
\usage{
parse_s3_uri(uri)
}
\arguments{
\item{uri}{An AWS S3 URI}
}
\value{
A named list with bucket and object elements
}
\description{
Parse an S3 URI
}
\examples{
parse_s3_uri("s3://bucket-name/a/path/to/an/object.txt")
}
