% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gam_splines}
\alias{plot_gam_splines}
\title{Plot GAM splines for each group}
\usage{
plot_gam_splines(gam_model, tract, df_tract, dwi_metric, out_dir)
}
\arguments{
\item{gam_model}{GAM object, produced by gam/bam}

\item{tract}{AFQ tract name}

\item{df_tract}{A dataframe of AFQ nodes for certain tract}

\item{dwi_metric}{Diffusion MRI metric (e.g. FA, MD)}

\item{out_dir}{directory in which to save plots}
}
\description{
Plot GAM splines for each group
}
\examples{
\dontrun{
df_afq <- read.csv("/path/to/afq/output.csv")
gam_fit <- fit_gam(df_afq,
                   target = "dti_fa",
                   covariates = list("group", "sex"),
                   family = "gamma",
                   k = 40)
plot_gam_splines(gam_fit, "OR", df_afq, "dti_fa", ".")
}
}
