% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tract_profiles}
\alias{plot_tract_profiles}
\title{Plot tract profiles for each bundle as a facet and each metric as a figure.}
\usage{
plot_tract_profiles(
  df,
  metrics = NULL,
  bundles = NULL,
  bundles_col = "tractID",
  group_col = NULL,
  line_func = "mean",
  linewidth = 1,
  ribbon_func = "mean_cl_boot",
  ribbon_alpha = 0.25,
  n_groups = 3,
  pal_name = "colorblind",
  out_dir = getwd(),
  figsize = c(8, 11.5)
)
}
\arguments{
\item{df}{Data frame.}

\item{metrics}{Name(s) of the metrics to plot per figure, character vector.
By default, will be all diffusion metrics in the provided data frame.}

\item{bundles}{Name(s) of the tract bundles to plot per facet, character
vector. By default, will be all tract bundles in the provided data
frame.}

\item{bundles_col}{Name of the column in the provided data frame with the
tract bundles.}

\item{group_col}{Name of the column in the data frame to group by as a color,
character. By default, no grouping variable is provided.}

\item{line_func}{Line function that provides the line positioning. See
\link[ggplot2]{stat_summary} for more information.}

\item{linewidth}{Line thickness of the tract profile line.}

\item{ribbon_func}{Ribbon function that provides the range for the ribbon.
See \link[ggplot2]{stat_summary} for more information.}

\item{n_groups}{Number of groups to}

\item{pal_name}{Grouping color palette name, character. Default is colorblind.}

\item{out_dir}{Output directory of saved plots.}

\item{figsize}{Figure size. A numeric vector of (width, height) in inches.}
}
\description{
Plot tract profiles for each bundle as a facet and each metric as a figure.
}
