% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read.afq.files}
\alias{read.afq.files}
\title{Create a merged AFQ/phenotype dataframe}
\usage{
read.afq.files(
  nodes_csv,
  pheno_csv,
  index = "subjectID",
  index.nodes = index,
  index.pheno = index,
  dwi_metrics = NULL,
  factor_cols = NULL,
  pheno_cols = NULL,
  ...
)
}
\arguments{
\item{nodes_csv}{path to a nodes file}

\item{pheno_csv}{path to a phenotypic file}

\item{index}{specification of the column used for merging}

\item{index.nodes}{specification of the column used for merging}

\item{index.pheno}{specification of the column used for merging}

\item{dwi_metrics}{which diffusion metrics should be retained}

\item{factor_cols}{which columns should be treated as factors}

\item{pheno_cols}{which columns to include from pheno file}

\item{...}{arguments to be passed to read.csv}
}
\description{
Create a merged AFQ/phenotype dataframe
}
\examples{
\dontrun{
  df_afq <- read.afq.files(nodes_csv = "a/path/to/nodes.csv",
                           pheno_csv = "a/path/to/pheno.csv")
}
}
