% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read.afq.weston.havens}
\alias{read.afq.weston.havens}
\title{Load tract profiles from Yeatman et al.}
\usage{
read.afq.weston.havens(...)
}
\arguments{
\item{...}{arguments to be passed to read.csv}
}
\value{
A merged dataframe with data from Yeatman et al.
}
\description{
Load tract profiles from Yeatman et al.
}
\examples{
df_weston_havens <- read.afq.weston.havens()
}
