% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_afq_files}
\alias{read_afq_files}
\title{Create a merged AFQ and phenotype dataframe}
\usage{
read_afq_files(
  nodes_file,
  pheno_file = NULL,
  by = "subjectID",
  tract_col = "tractID",
  node_col = "nodeID",
  other_cols = NULL,
  na_omit = FALSE,
  ...
)
}
\arguments{
\item{nodes_file}{Path to a nodes file.}

\item{pheno_file}{Path to a phenotypic file. Leave NULL to return an
"unsupervised" AFQ dataset}

\item{by}{Column used for merging the nodes and phenotypic files.
If variable names differ between the nodes and
phenotypic datasets, use a named character vector like
by = c("x_a" = "y_a", "x_b" = "y_b").
Default: "subjectID"}

\item{tract_col}{description}

\item{node_col}{description}

\item{other_cols}{description}

\item{na_omit}{description}

\item{...}{}
}
\description{
There should a description of what an "AFQ" dataset is
}
\examples{
\dontrun{
df_afq <- read_afq_files(
  nodes_file = "path/to/nodes.csv",
  pheno_file = "path/to/pheno.csv"
)
}
}
