% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_afq_hbn}
\alias{read_afq_hbn}
\title{Load tract profiles from the Healthy Brain Network dataset}
\usage{
read_afq_hbn(truncate = FALSE, na_omit = TRUE, ...)
}
\arguments{
\item{truncate}{Truncate the data to 49 rows, boolean.
Default: FALSE.}

\item{...}{Arguments to be passed to read.csv}
}
\value{
A merged dataframe with data from HBN
}
\description{
Load tract profiles from the Healthy Brain Network dataset
}
\examples{
\dontrun{
  df_hbn <- read_afq_hbn()
}
}
