% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_afq_sarica}
\alias{read_afq_sarica}
\title{Load tract profiles from Sarica et al.}
\usage{
read_afq_sarica(na_omit = FALSE, ...)
}
\arguments{
\item{na_omit}{A logical indicating whether rows with NA values should be
removed from the data frame before returning.
Default: FALSE}

\item{...}{Further keyword arguments to be passed to the file reading
function: \link[readr:read_delim]{readr::read_csv} or \link[readr:read_delim]{readr::read_tsv}.}
}
\value{
A merged dataframe with data from Sarica et al.
}
\description{
Load tract profiles from Sarica et al.
}
\examples{
df_sarica <- read_afq_sarica()
}
