% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{save_gam}
\alias{save_gam}
\title{DESCRIPTION}
\usage{
save_gam(
  gam_model,
  file,
  model_rdata = TRUE,
  model_summary = TRUE,
  model_check = FALSE
)
}
\arguments{
\item{model_check}{}
}
\value{
None.
}
\description{
A short description...
}
