% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_bundle}
\alias{select_bundle}
\title{Select a single bundle from an AFQ dataframe}
\usage{
select_bundle(
  df_afq,
  tract,
  dwi_metric,
  covariates,
  participant_id = "subjectID",
  group_by = "group"
)
}
\arguments{
\item{df_afq}{The input AFQ dataframe}

\item{tract}{Abbreviated tract name, e.g., "CST_L" or "OR"}

\item{dwi_metric}{The diffusion MRI metric (e.g. "FA", "MD")}

\item{covariates}{List of strings of GAM covariates, not including
the smoothing terms over nodes and the random effect due to subjectID.
This list can also include smoothing terms.}

\item{participant_id}{The name of the column that encodes participant ID}

\item{group_by}{The grouping variable used to group nodeID smoothing terms}
}
\value{
An AFQ dataframe containing only the selected bundle
}
\description{
Select a single bundle from an AFQ dataframe
}
