% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle_df}
\alias{shuffle_df}
\title{Shuffle an AFQ dataframe}
\usage{
shuffle_df(input_df, dwi_metric)
}
\arguments{
\item{input_df}{The input AFQ dataframe}

\item{dwi_metric}{The diffusion MRI metric (e.g. "FA", "MD")}
}
\value{
A shuffled AFQ dataframe
}
\description{
This function shuffles participants' age, group, and sex,
thereby destroying correlations between the participants'
tract profiles and phenotypic data.
}
\examples{
\dontrun{
df_afq <- read.csv("/path/to/afq/output.csv")
df_shuffle <- shuffle_df(df_afq, "dti_fa")
}
}
