% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{spline_diff}
\alias{spline_diff}
\title{Calculate and plot difference between two splines}
\usage{
spline_diff(
  gam_model,
  tract,
  group_by = "group",
  factor_a,
  factor_b,
  save_output = TRUE,
  sim.ci = FALSE,
  out_dir
)
}
\arguments{
\item{gam_model}{GAM object, produced by gam/bam}

\item{tract}{AFQ tract name}

\item{group_by}{The grouping variable used to group nodeID smoothing terms}

\item{factor_a}{First group factor, string}

\item{factor_b}{Second group factor, string}

\item{save_output}{Boolean. If TRUE, save plot output}

\item{sim.ci}{Logical: Using simultaneous confidence intervals or not
(default set to FALSE). The implementation of simultaneous CIs follows
Gavin Simpson's blog of December 15, 2016:
http://www.fromthebottomoftheheap.net/2016/12/15/simultaneous-interval-revisited/.
This interval is calculated from simulations based. Please specify a seed
(e.g., set.seed(123)) for reproducable results. Note: in contrast with
Gavin Simpson's code, here the Bayesian posterior covariance matrix of
the parameters is uncertainty corrected (unconditional=TRUE) to reflect
the uncertainty on the estimation of smoothness parameters.}

\item{out_dir}{Directory in which to save plots}
}
\value{
A dataframe of spline differences at each node
}
\description{
This function both
\enumerate{
\item draws a spline-difference plot for 2 splines,
\item and returns a dataframe of differences at each node
}
}
\examples{
\dontrun{
df_afq <- read.csv("/path/to/afq/output.csv")
gam_fit <- fit_gam(df_afq,
                   target = "dti_fa",
                   covariates = list("group", "sex"),
                   family = "gamma",
                   k = 40)
df_diff = spline_diff(gam_model = gam_fit,
                      tract = "OR",
                      factor_a = "0",
                      factor_b = "1",
                      out_dir = ".")
}
}
