% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tractable-package.R
\docType{package}
\name{tractable-package}
\alias{tractable}
\alias{tractable-package}
\title{tractable: Analyzing diffusion MRI tract profiles}
\description{
A toolbox for analyzing diffusion MRI tract profiles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tractometry/tractable}
  \item Report bugs at \url{https://github.com/tractometry/tractable/issues}
}

}
\author{
\strong{Maintainer}: Adam Richie-Halford \email{richiehalford@gmail.com} (\href{https://orcid.org/0000-0001-9276-9084}{ORCID})

Authors:
\itemize{
  \item Kelly Chang \email{kchang4@uw.edu} (\href{https://orcid.org/0000-0001-5385-2782}{ORCID})
  \item Teresa Gomez \email{tmgomez@uw.edu} (\href{https://orcid.org/XXX}{ORCID})
  \item McKenzie Hagen \email{mphagen@uw.edu} (\href{https://orcid.org/0000-0002-7454-8189}{ORCID})
  \item Ethan Roy \email{ethanroy395@gmail.com} (\href{https://orcid.org/0000-0002-4956-2237}{ORCID})
  \item Ariel Rokem \email{arokem@gmail.com} (\href{https://orcid.org/0000-0003-0679-1985}{ORCID})
}

}
\keyword{internal}
