% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tractable.R
\name{tractable_single_tract}
\alias{tractable_single_tract}
\title{Analyze group differences in a single dMRI tract profile using GAMs}
\usage{
tractable_single_tract(df, tract, tract_col = "tractID", ...)
}
\arguments{
\item{df}{Input AFQ dataframe. If NULL, this function will load data
using read_afq_files and the additional arguments in ...}

\item{tract}{Abbreviated tract name, e.g., "CST_L" or "OR"}

\item{tract_col}{The diffusion metric to model (e.g. "FA", "MD")}

\item{...}{Further keyword arguments to be passed to
\link{fit_gam}.}
}
\description{
Analyze group differences in a single dMRI tract profile using GAMs
}
\examples{
df_sarica <- read_afq_sarica()

tractable_single_tract(
  df         = df_sarica, 
  tract      = "Right Corticospinal",
  target     = "fa",
  regressors = c("age", "group"),
  node_group = "group"
)
}
