% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{bootstrap_df}
\alias{bootstrap_df}
\title{Bootstrap an AFQ dataframe}
\usage{
bootstrap_df(
  df,
  target,
  node_col = "nodeID",
  node_group = "group",
  participant_col = "subjectID"
)
}
\arguments{
\item{df}{The input dataframe.}

\item{target}{The column name that encodes the metric to model.}

\item{node_col}{The column name that encodes tract node positions.
Default: "nodeID"}

\item{node_group}{The column name to group the tract node smooth by.
Default: NULL.}

\item{participant_col}{The column name that encodes participant ID.
Default: "subjectID".}
}
\value{
A shuffled AFQ dataframe
}
\description{
This function bootstrap samples an AFQ dataframe by participant. That is, it
first pivots to wide format with one row per participant, bootstrap samples,
and finally pivots back to long format.
}
\examples{
\dontrun{
df_afq <- read.csv("/path/to/afq/output.csv")
df_bootstrap <- bootstrap_df(df_afq, "dti_fa")}
}
