% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{build_formula}
\alias{build_formula}
\title{Build a GAM formula.}
\usage{
build_formula(
  target,
  regressors = NULL,
  node_k = 3,
  node_col = "nodeID",
  participant_col = "subjectID"
)
}
\arguments{
\item{target}{The column name that encodes the metric to model.}

\item{regressors}{Column name or list of column names to use as
regressors, not including nodes smoothing terms and
the participant random effect. This list can also
include smoothing terms. Default: NULL.}

\item{node_k}{The basis dimensions used to represent the node
smoother. Default: 3.}

\item{node_col}{The name of the column that encode node ID.
Default: "nodeID".}

\item{participant_col}{The name of the column that encodes participant ID.
Default: "participantID"}
}
\value{
A GAM formula string.
}
\description{
Function used to build a generic GAM formula string from the given arguments.
The function automatically includes a tract node smoothing term and a
participant random effects (random intercept). Any regressor terms are
inlcuded as additive effects.
}
\examples{
formula <- build_formula(target = "dti_fa",
                         regressors = c("group", "sex"),
                         node_k = 40)

formula <- build_formula(target = "dki_md",
                         regressors = c("group", "sex", "s(age, by = sex)"),
                         node_k = 32)
}
