% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{build_formula}
\alias{build_formula}
\title{Build a GAM formula.}
\usage{
build_formula(
  target,
  regressors = NULL,
  node_col = "nodeID",
  node_k = 10,
  node_group = NULL,
  participant_col = "subjectID"
)
}
\arguments{
\item{target}{The column name that encodes the metric to model.}

\item{regressors}{Column name or list of column names to use as
regressors, not including nodes smoothing terms and
the participant random effect. This list can also
include smoothing terms. Default: NULL.}

\item{node_col}{The column name that encodes tract node positions.
Default: "nodeID".}

\item{node_k}{The basis dimensions used to represent the node
smoother. If \code{node_group}, the basis value is applied
to the group as well. Default: 10.}

\item{node_group}{The column name to group the tract node smooth by.
Default: NULL.}

\item{participant_col}{The column name that encodes participant ID.
Default: "subjectID".}
}
\value{
A GAM formula.
}
\description{
Function used to build a generic GAM formula string from the given arguments.
The function automatically includes a tract node smoothing term (with or
without a grouping term) and a participant random effects (random intercept).
Any regressor terms are included as additive effects.
}
\examples{
formula <- build_formula(target = "dti_fa", node_k = 40)

formula <- build_formula(target = "dki_md",
                         regressors = c("group", "sex"), 
                         node_k = 32, 
                         node_group = "group")
}
