% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{estimate_distribution}
\alias{estimate_distribution}
\title{Estimate distribution function from values.}
\usage{
estimate_distribution(
  x,
  distr_options = c("beta", "gamma", "gaussian"),
  eval_metric = "aic"
)
}
\arguments{
\item{x}{A numeric vector that will be evaluated.}

\item{eval_metric}{The distribution evaluation metric names. \cr
Possible options: ("aic", "bic", "loglik") \cr
Default: "aic"}

\item{distr_names}{A vector of distribution names to evaluate the values. \cr
Possible options: ("beta", "gamma", "gaussian") \cr
Default: c("beta", "gamma", "gaussian")}
}
\value{
The estimated \code{family} or \code{extended.family} function.
}
\description{
Estimate distribution function from values.
}
\examples{
x <- rnorm(1000)
estimate_distribution(x)

x <- rgamma(1000, shape = 12)
estimate_distribution(x)
}
\seealso{
\link[fitdistrplus:fitdist]{fitdistrplus::fitdist}
}
