% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{fit_gam}
\alias{fit_gam}
\alias{fit_gam.formula}
\title{Fit a GAM model.}
\usage{
fit_gam(...)

\method{fit_gam}{formula}(
  formula,
  df,
  node_col = "nodeID",
  autocor = TRUE,
  family = "auto",
  method = "fREML",
  discrete = TRUE,
  ...
)
}
\arguments{
\item{formula}{Explicit formula to use for the GAM.}

\item{df}{The data frame contains GAM metrics.}

\item{node_col}{The column name that encodes tract node positions.
Default: "nodeID".}

\item{target}{The column name that encodes the metric to model.}

\item{regressors}{Column name or list of column names to use as
regressors, not including nodes smoothing terms and
the participant random effect. This list can also
include smoothing terms. Default: NULL.}

\item{node_k}{The basis dimensions used to represent the node
smoother. If \code{node_group}, the basis value is applied
to the group as well. Default: 'auto'.}

\item{node_group}{The column name to group the tract node smooth by
(i.e., \code{s(node_col, by = node_group, k = node_k)}).}
}
\value{
Fit GAM model
}
\description{
Function used to fit a GAM formula from an explicit formula or builds a
formula from the given arguments.
}
\details{
This function has a series of steps:
\enumerate{
\item If family == "auto", choose the distribution (either 'beta', 'gamma',
or 'norm') that has the lowest AIC when fitting to the GAM dependent
variable data using \code{\link[fitdistrplus:fitdist]{fitdistrplus::fitdist()}}.
\item If node_k == "auto", build an initial GAM model with k = 2 and
continue to double the k value until gam.check shows that k is
large enough.
\item Fit a GAM model according to the formula given or build a formula
that follow the below pattern:
\cr \cr
target ~ regressor1 (+ regressor2 + ...) +
s(node_col, by = node_group, k = node_k) +
s(participant_col, bs = "re")
\cr
\item Optionally, save the output of gam.check and summary to files.
}
}
\examples{
\dontrun{
df_afq <- read.csv("/path/to/afq/output.csv")

df_tract <- df_afq \%>\% 
  filter(tractID == "CST_L")

gam_fit <- fit_gam(
  target = "dti_fa", 
  regressors = c("sex", "group"), 
  df = df_tract, 
  family = "gamma", 
  node_k = "auto"
)
}
}
