% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{permutation_test}
\alias{permutation_test}
\title{Simulate the null distribution using permutation testing}
\usage{
permutation_test(
  df_tract,
  n_permutations,
  dwi_metric,
  covariates = NULL,
  k = NULL,
  family = NULL,
  formula = NULL
)
}
\arguments{
\item{df_tract}{AFQ Dataframe of node metric values for single tract}

\item{n_permutations}{Number of permutation tests to perform}

\item{dwi_metric}{The diffusion metric to model (e.g. "FA", "MD")}

\item{covariates}{List of strings of GAM covariates, not including
the smoothing terms over nodes and the random effect due to subjectID.
This list can also include smoothing terms.}

\item{k}{Dimension of the basis used to represent the node smoothing term}

\item{family}{Distribution to use for the gam. Must be 'gamma' or 'beta'}

\item{formula}{Optional explicit formula to use for the GAM. If provided,
this will override the dynamically generated formula build from the
target, covariate, and k inputs. Default = NULL.}
}
\value{
Dataframe with permutation test coefficients
}
\description{
Simulate the null distribution using permutation testing
}
\examples{
\dontrun{
df_afq <- read.csv("/path/to/afq/output.csv")
df_tract <- df_afq[which(df_afq$tractID == tract), ]
permutation_coefs <- fit_gam(df_afq,
                             dwi_metric = "dti_fa",
                             covariates = list("group", "sex"),
                             family = "gamma",
                             k = 40,
                             n_permutations = 1000)
}
}
