% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gam_splines}
\alias{plot_gam_splines}
\title{Plot GAM splines for each group}
\usage{
plot_gam_splines(
  gam_model,
  tract,
  df_tract,
  dwi_metric,
  covariates,
  group_by = "group",
  participant_id = "subjectID",
  out_dir
)
}
\arguments{
\item{gam_model}{GAM object, produced by gam/bam}

\item{tract}{AFQ tract name}

\item{df_tract}{A dataframe of AFQ nodes for certain tract}

\item{dwi_metric}{Diffusion MRI metric (e.g. FA, MD)}

\item{covariates}{List of strings of GAM covariates,
not including the smoothing terms over nodes and the random effect due
to subjectID.}

\item{group_by}{The grouping variable used to group nodeID smoothing terms}

\item{participant_id}{The name of the column that encodes participant ID}

\item{out_dir}{directory in which to save plots}
}
\description{
Plot GAM splines for each group
}
\examples{
\dontrun{
df_afq <- read.csv("/path/to/afq/output.csv")
gam_fit <- fit_gam(df_afq,
                   target = "dti_fa",
                   covariates = list("group", "sex"),
                   family = "gamma",
                   k = 40)
plot_gam_splines(gam_model = gam_fit,
                 tract = "OR",
                 df_tract = df_afq,
                 dwi_metric = "dti_fa",
                 covariates = c("group", "sex"),
                 out_dir = ".")
}
}
