% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tract_profiles}
\alias{plot_tract_profiles}
\title{Plot tract profiles for each tract as a facet and each metric as a figure.}
\usage{
plot_tract_profiles(
  df,
  y = NULL,
  tracts = NULL,
  tract_col = "tractID",
  group_col = NULL,
  line_func = "mean",
  linewidth = 1,
  ribbon_func = "mean_cl_boot",
  ribbon_alpha = 0.25,
  n_groups = 3,
  pal_name = "colorblind",
  output_dir = getwd(),
  ...
)
}
\arguments{
\item{df}{Data frame.}

\item{tracts}{Name(s) of the tract tracts to plot per facet, character
vector. By default, will be all tract tracts in the provided data
frame.}

\item{tract_col}{Name of the column in the provided data frame with the
tracts.}

\item{group_col}{Name of the column in the data frame to group by as a
color, character. By default, no grouping variable is provided.}

\item{line_func}{Line function that provides the line positioning. See
\link[ggplot2:stat_summary]{ggplot2::stat_summary} for more information.}

\item{linewidth}{Line thickness of the tract profile line.}

\item{ribbon_func}{Ribbon function that provides the range for the ribbon.
See \link[ggplot2:stat_summary]{ggplot2::stat_summary} for more information.}

\item{ribbon_alpha}{Ribbon alpha level.}

\item{n_groups}{Number of groups to split a numeric grouping variable.}

\item{pal_name}{Grouping color palette name, character.
Default: Colorblind palette.}

\item{output_dir}{Output directory for the plot images.
Default: Current working directory.}

\item{...}{Keyword arguments to be passed to \link[ggplot2:ggsave]{ggplot2::ggsave}.}

\item{metrics}{Name(s) of the metrics to plot per figure, character vector.
By default, will be all diffusion metrics in the
provided data frame.}
}
\value{
List of plot handles corresponding to the specified metrics.
}
\description{
Plot tract profiles for each tract as a facet and each metric as a figure.
}
\examples{
\dontrun{
df_sarica <- read_afq_sarica(na_omit = TRUE)

plot_tract_profiles(
  df_sarica,
  metrics = c("fa"),
  tracts = c("Left Corticospinal", "Right Corticospinal"),
  group_col = "class",
  figsize   = c(8, 6)
)

plot_tract_profiles(
  df_sarica,
  metrics   = "fa"
  tracts    = c("Left Corticospinal", "Right Corticospinal"),
  group_col = "age",
  n_groups  = 3,
  pal_name  = "Spectral",
  figsize   = c(8, 6)
)
}
}
