% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tract_profiles}
\alias{plot_tract_profiles}
\title{Plot tract profiles}
\usage{
plot_tract_profiles(
  df,
  y,
  tracts = NULL,
  tract_col = "tractID",
  node_col = "nodeID",
  group_col = NULL,
  n_groups = NULL,
  group_pal = "cb",
  participant_col = "subjectID",
  ribbon_func = "mean_cl_boot",
  ribbon_alpha = 0.25,
  linewidth = 1,
  save_figure = TRUE,
  output_dir = getwd(),
  ...
)
}
\arguments{
\item{df}{The input dataframe.}

\item{y}{Column name(s) of the variables to plot on the y-axis.}

\item{tracts}{Name(s) of the tract tracts to plot per facet.
Default: NULL \cr \cr
If NULL, will be all tracts in the data frame.}

\item{tract_col}{The column name that encodes the tract information.
Default: tractID}

\item{node_col}{The column name that encodes tract node positions.
Default: "nodeID".}

\item{group_col}{Column name that encodes group information. Will be
drawn by color. Default: NULL.
\itemize{
\item  If \code{group_col} is a factor, will use unique
values as groups.
\item  If \code{group_col} is a numeric, will use
\link[ggplot2:cut_interval]{ggplot2::cut_interval} to create equal range
\code{n_groups} as groups.
}}

\item{n_groups}{Number of groups to split a numeric grouping variable.
Only used if \code{group_col} is numeric.
Default: NULL.}

\item{group_pal}{Grouping color palette name. Valid options include:
"cb" (colorblind), "cbb" (colorblind_black), and all
named \link{RColorBrewer} palettes.
Default: "cb" (colorblind)}

\item{participant_col}{The column name that encodes participant ID.
Default: "subjectID".}

\item{ribbon_func}{Ribbon summarizing function that provides the y, ymin,
and ymax for the ribbon. See \code{fun.data} argument from
\link[ggplot2:stat_summary]{ggplot2::stat_summary} for more information.
Default: "mean_cl_boot"}

\item{ribbon_alpha}{Ribbon alpha level. Default: 0.25}

\item{linewidth}{Line thickness of the tract profile line. Default: 1.}

\item{save_figure}{Boolean flag. If TRUE, save tract profiles. If FALSE,
do not save tract profiles. Default: FALSE}

\item{output_dir}{Output directory for the figure image(s).
Default: \link{getwd} (current working directory).}

\item{...}{Keyword arguments to be passed to \link[ggplot2:ggsave]{ggplot2::ggsave}.}
}
\value{
Named list of plot handles corresponding to the specified y values.
}
\description{
Create tract profile figures for each tract as a facet and for each \code{y}
value as a figure.
}
\details{
If \code{save_figure} is TRUE, the naming convention is as follows:
\itemize{
\item If \code{group_col}, "tract_by-(group_col)_param-(y)_profiles.png".
\item If \code{group_col == NULL}, "tract_param-(y)_profiles.png"
}
}
\examples{
\dontrun{
df_sarica <- read_afq_sarica(na_omit = TRUE)

plot_handle <- plot_tract_profiles(
  df        = df,
  y         = c("fa", "md"),
  tracts    = c("Left Corticospinal", "Right Corticospinal"),
  width     = 12, 
  height    = 6,
  units     = "in"
)

plot_handle <- plot_tract_profiles(
  df        = df,
  y         = c("fa", "md"),
  tracts    = c("Left Corticospinal", "Right Corticospinal"),
  group_col = "group", 
  width     = 12, 
  height    = 6,
  units     = "in"
)

plot_handle <- plot_tract_profiles(
  df        = df,
  y         = "fa",
  tracts    = c("Left Corticospinal", "Right Corticospinal"),
  group_col = "age", 
  n_groups  = 5, 
  group_pal = "Spectral", 
  width     = 12, 
  height    = 6,
  units     = "in"
)}
}
