% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_afq_files}
\alias{read_afq_files}
\title{Create a merged AFQ and phenotype dataframe}
\usage{
read_afq_files(
  nodes_file,
  pheno_file = NULL,
  by = "subjectID",
  tract_col = "tractID",
  node_col = "nodeID",
  other_cols = NULL,
  na_omit = FALSE,
  ...
)
}
\arguments{
\item{nodes_file}{Path to a nodes file.}

\item{pheno_file}{Path to a phenotypic file. If pheno_file = NULL
returns an "unsupervised" AFQ dataset.}

\item{by}{Column used for merging the nodes and phenotypic files.
Default: "subjectID"
\cr \cr
If variable names differ between the nodes and
phenotypic datasets, use a named character vector like
by = c("x_a" = "y_a", "x_b" = "y_b").}

\item{tract_col}{The column name that encodes tract names.
Default: "tractID"}

\item{node_col}{The column name that encodes tract node positions.}

\item{other_cols}{Other column names to be included in the final
data frame. Default: NULL}

\item{na_omit}{A logical indicating whether rows with NA values should
be removed from the data frame before returning.
Default: FALSE}

\item{...}{Further keyword arguments to be passed to the file
reading function (\link[readr:read_delim]{readr::read_csv}, \link[readr:read_delim]{readr::read_tsv})}
}
\value{
An AFQ dataset.
}
\description{
Create a merged AFQ and phenotype dataframe
}
\examples{
\dontrun{
df_afq <- read_afq_files(
  nodes_file = "path/to/nodes.csv",
  pheno_file = "path/to/pheno.csv"
)
}
}
