% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_afq_files}
\alias{read_afq_files}
\title{Create a merged AFQ/phenotype dataframe}
\usage{
read_afq_files(
  nodes_csv,
  pheno_csv = NULL,
  index = "subjectID",
  index_nodes = index,
  index_pheno = index,
  dwi_metrics = NULL,
  factor_cols = NULL,
  pheno_cols = NULL,
  ...
)
}
\arguments{
\item{nodes_csv}{path to a nodes file}

\item{pheno_csv}{path to a phenotypic file, leave NULL to return an "unsupervised" AFQ dataset}

\item{index}{specification of the column used for merging}

\item{index_nodes}{specification of the column used for merging}

\item{index_pheno}{specification of the column used for merging}

\item{dwi_metrics}{which diffusion metrics should be retained}

\item{factor_cols}{which columns should be treated as factors}

\item{pheno_cols}{which columns to include from pheno file}

\item{...}{arguments to be passed to read.csv}
}
\description{
Create a merged AFQ/phenotype dataframe
}
\examples{
\dontrun{
  df_afq <- read_afq_files(nodes_csv = "a/path/to/nodes.csv",
                           pheno_csv = "a/path/to/pheno.csv")
}
}
