% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_afq_sarica}
\alias{read_afq_sarica}
\title{Load tract profiles from Sarica et al.}
\usage{
read_afq_sarica(na_omit = TRUE, ...)
}
\arguments{
\item{...}{arguments to be passed to read.csv}
}
\value{
A merged dataframe with data from Sarica et al.
}
\description{
Load tract profiles from Sarica et al.
}
\examples{
df_sarica <- read_afq_sarica()
}
