% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{read_afq_weston_havens}
\alias{read_afq_weston_havens}
\title{Load tract profiles from Yeatman et al.}
\usage{
read_afq_weston_havens(na_omit = FALSE, ...)
}
\arguments{
\item{na_omit}{A logical indicating whether rows with NA values should be
removed from the data frame before returning.
Default: FALSE}

\item{...}{Further keyword arguments to be passed to the file reading
function (\link[readr:read_delim]{readr::read_csv}, \link[readr:read_delim]{readr::read_tsv}).}
}
\value{
A merged dataframe with data from Yeatman et al.
}
\description{
Load tract profiles from Yeatman et al.
}
\examples{
df_weston_havens <- read_afq_weston_havens()
}
