% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{save_gam}
\alias{save_gam}
\title{Save GAM model outputs.}
\usage{
save_gam(gam_model, output_file, model_summary = TRUE, model_check = FALSE)
}
\arguments{
\item{gam_model}{GAM model object to be saved as an RData file.}

\item{output_file}{RData output file name. Accepted a file extensions:
.RData, .Rdata, .rdata, or .rda.}

\item{model_summary}{A logical indicating whether the model summary output
should be saved as a text file.
Default: TRUE}

\item{model_check}{A logical indicating whether the model check reports
should be saved.
Default: FALSE}
}
\description{
This function saves the GAM model as RData (.rda) file. Optionally, it will
save the model summary as a text (.txt) file, the \link[mgcv:gam.check]{mgcv::gam.check}
figures as a PNG (.png), and the \link[mgcv:k.check]{mgcv::k.check} output as a text (.txt) file.
}
\details{
If \code{model_summary} is TRUE, the output of \code{summary(gam_model)} will
be captured and saved in a text file with the following naming convention
"<output_file>_Summary.txt".

If \code{model_check} is TRUE, the outputs of \link[mgcv:gam.check]{mgcv::gam.check} will be
captured and saved. \link[mgcv:gam.check]{mgcv::gam.check} consists of two components:
\enumerate{
\item GAM model diagnostic plots, implemented through \link[gratia:appraise]{gratia::appraise}. The
output will be saved with following naming convention:
"<output_file>_GamCheck.png"
\item Smoothing basis dimension checks, implemented through \link[mgcv:k.check]{mgcv::k.check}. The
output will be saved with the following naming convention:
"<output_file>_GamCheck.txt"
}
}
\examples{
\dontrun{
save_gam(
  gam_model   = model_fit, 
  output_file = "output.rda"
)}
}
