% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{save_gam}
\alias{save_gam}
\title{Save GAM model outputs.}
\usage{
save_gam(
  gam_model,
  file,
  model_rdata = TRUE,
  model_summary = TRUE,
  model_check = FALSE
)
}
\arguments{
\item{gam_model}{asdfasdf}

\item{file}{asdfasdf}

\item{model_rdata}{asdfasdf}

\item{model_summary}{asdfasdf}

\item{model_check}{asdfasdf}
}
\value{
None.
}
\description{
This function saves the GAM model as RData (.rda) file, the model summary as
a text (.txt) file, and the \code{\link[mgcv:gam.check]{mgcv::gam.check()}} figures as a PNG (.png) file.
}
