% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{shuffle_df}
\alias{shuffle_df}
\title{Shuffle an AFQ dataframe}
\usage{
shuffle_df(
  df,
  target,
  shuffle_cols = NULL,
  node_col = "nodeID",
  node_group = NULL,
  tract_col = "tractID",
  participant_col = "subjectID",
  sample_uniform = FALSE
)
}
\arguments{
\item{df}{The input AFQ dataframe}

\item{target}{The diffusion MRI metric (e.g. "FA", "MD")}

\item{shuffle_cols}{Column names that should be shuffled.}

\item{node_col}{sdecriptiojn}

\item{node_group}{The grouping variable used to group nodeID smoothing terms}

\item{tract_col}{description}

\item{participant_col}{The name of the column that encodes participant ID}

\item{sample_uniform}{Boolean flag. If TRUE, shuffling should sample
uniformly from the unique values in the columns. If FALSE, shuffling
will shuffle without replacement.}
}
\value{
A shuffled AFQ dataframe
}
\description{
This function shuffles participants' age, group, and sex,
thereby destroying correlations between the participants'
tract profiles and phenotypic data.
}
\examples{
\dontrun{
df_afq <- read_csv("/path/to/afq/output.csv")
df_shuffle <- shuffle_df(df_afq, target = "dti_fa")
}
}
