% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_diff.R
\name{spline_diff}
\alias{spline_diff}
\title{Calculate and plot difference between two splines}
\usage{
spline_diff(gam_model, tract, group.by = "group", factor_a, factor_b, out_dir)
}
\arguments{
\item{gam_model}{GAM object, produced by gam/bam}

\item{tract}{AFQ tract name}

\item{group.by}{The grouping variable used to group nodeID smoothing terms}

\item{factor_a}{First group factor, string}

\item{factor_b}{Second group factor, string}

\item{out_dir}{Directory in which to save plots}
}
\value{
A dataframe of spline differences at each node
}
\description{
This function both
\enumerate{
\item draws a spline-difference plot for 2 splines,
\item and returns a dataframe of differences at each node
}
}
\examples{
\dontrun{
df_afq <- read.csv("/path/to/afq/output.csv")
gam_fit <- fit_gam(df_afq,
                   target = "dti_fa",
                   covariates = list("group", "sex"),
                   family = "gamma",
                   k = 40)
df_diff = spline_diff(gam_model = gam_fit,
                      tract = "OR",
                      factor_a = "0",
                      factor_b = "1",
                      out_dir = ".")
}
}
