% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tract_name}
\alias{tract_name}
\title{Retrieve tract name from abbreviation}
\usage{
tract_name(tract_abbr)
}
\arguments{
\item{tract_abbr}{AFQ tract abbreviation}
}
\value{
Formatted tract name
}
\description{
Retrieve tract name from abbreviation
}
\examples{
full_name <- tract_name("OR")
full_name <- tract_name("CST_L")
}
