% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tractable.R
\name{tractable_single_tract}
\alias{tractable_single_tract}
\title{Analyze group differences in a single dMRI tract profile using GAMs}
\usage{
tractable_single_tract(df, tract, tract_col = "tractID", ...)
}
\arguments{
\item{df}{Input data frame.}

\item{tract}{Tract name to subset from the dataframe.}

\item{tract_col}{The column name that encodes the tract information.
Default: tractID}

\item{...}{
  Arguments passed on to \code{\link[=fit_gam]{fit_gam}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
GAM model fit object.
}
\description{
Analyze group differences in a single dMRI tract profile using GAMs
}
\examples{
df_sarica <- read_afq_sarica(na_omit = TRUE)

tractable_single_tract(
  df         = df_sarica, 
  tract      = "Right Corticospinal",
  target     = "fa",
  regressors = c("age", "group"),
  node_group = "group"
)
}
