% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tractr}
\alias{tractr}
\title{Analyze group differences in dMRI tract profiles data using GAMs}
\usage{
tractr(
  afq.csv,
  tract,
  dwi_metric,
  covariates,
  comp_list,
  out_dir,
  permute.test = FALSE,
  permute.n = 100
)
}
\arguments{
\item{afq.csv}{Input CSV file with AFQ outputs}

\item{tract}{Abbreviated tract name, e.g., "CST_L" or "OR"}

\item{dwi_metric}{The diffusion metric to model (e.g. "FA", "MD")}

\item{covariates}{List of strings of GAM covariates,
not including the smoothing terms over nodes and the random effect due
to subjectID. This list can also include smoothing terms.}

\item{comp_list}{List of factor strings for pairwise comparison.
These should be values from the group column in the AFQ CSV file.}

\item{out_dir}{Directory in which to save text output and plots}

\item{permute.test}{If TRUE, estimate the null distribution of the overall
group coefficient using a permutation test.}

\item{permute.n}{Number of permutations to simulate the null distribution}
}
\description{
Analyze group differences in dMRI tract profiles data using GAMs
}
\examples{
\dontrun{
tractr(afq.csv = "/path/to/afq/output.csv",
       tract = "CST_R",
       dwi_metric = "dti_fa",
       covariates = c("sex", "group"),
       comp_list = c("0", "1"),
       out_dir = ".",
       permute.test = TRUE,
       permute.n = 100)
}
}
