% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tractr.R
\name{tractr_single_bundle}
\alias{tractr_single_bundle}
\title{Analyze group differences in a single dMRI tract profile using GAMs}
\usage{
tractr_single_bundle(
  df_afq,
  tract,
  dwi_metric,
  participant_id = "subjectID",
  group_by = "group",
  covariates = c(group_by),
  smooth_terms = NULL,
  k = "auto",
  family = "auto",
  ...
)
}
\arguments{
\item{df_afq}{Input AFQ dataframe. If NULL, this function will load data
using read.afq.data and the additional arguments in ...}

\item{tract}{Abbreviated tract name, e.g., "CST_L" or "OR"}

\item{dwi_metric}{The diffusion metric to model (e.g. "FA", "MD")}

\item{participant_id}{The name of the column that encodes participant ID}

\item{group_by}{The grouping variable used to group nodeID smoothing terms}

\item{covariates}{List of strings of GAM covariates,
not including the smoothing terms over nodes and the random effect due
to subjectID.}

\item{smooth_terms}{Smoothing terms, not including
the smoothing terms over nodes and the random effect due to subjectID.}

\item{k}{Dimension of the basis used to represent the node smoothing term,
If k = 'auto', this function will attempt to find the best value}

\item{family}{Distribution to use for the gam. Must be either 'gamma',
'beta', or 'auto'. If 'auto', this function will select the best fit
between beta and gamma distributions.}

\item{...}{Arguments to pass to fit_gam}
}
\description{
Analyze group differences in a single dMRI tract profile using GAMs
}
\examples{
sarica <- read.afq.sarica()
sarica$group <- factor(sarica$class)
sarica$subjectID <- unclass(factor(sarica$subjectID))
tractr_single_bundle(df_afq = sarica,
                     tract = "Right Corticospinal",
                     participant_id = "subjectID",
                     group_by = "group",
                     covariates = c("age","group"),
                     dwi_metric = "fa")
}
